updated city section and button
