+++ src/components/sections/AboutSection/AboutSection.css
@@ -1,6 +1,5 @@
 .about-kukdukoo{
   font-family: 'Creato Display', sans-serif;
-  /* font-weight: 175; */
   font-style: normal; 
   font-size: 20px;
   letter-spacing: 0; 
@@ -8,4 +7,14 @@
 }
 .about-kukdukoo strong {
   font-weight: 700;                      
+}
+.read-button {
+  margin-top: 20px;
+  padding: 10px 20px;
+  font-size: 16px;
+  color: white;
+  background-color: #007BFF;
+  border: none;
+  border-radius: 5px;
+  cursor: pointer;
 }
