+++ src/components/sections/CitySection/CitySection.css
@@ -1,0 +1,27 @@
+.location-icon{
+    width: 35px;
+    height: 35px;
+}
+.cities-section h2{
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  gap: 10px;
+  color: #000000;
+
+}
+.explore-city-button{
+  background-color: #d4ff65; 
+  border: 2px solid #000;
+  border-radius: 25px;
+  padding: 0.8rem 1.5rem;
+  font-weight: bold;
+  font-size: 1rem;
+  cursor: pointer;
+  box-shadow: 3px 3px 0 #000;
+  transition: all 0.2s ease;
+}
+.explore-city-button:hover{
+  background-color: #b6f052;
+  transform: scale(1.05);
+}
