+++ src/pages/HomePage.jsx
@@ -1,10 +1,10 @@
 import React from 'react';
 import HeroSection from '../components/sections/HeroSection/HeroSection';
 import AboutSection from '../components/sections/AboutSection/AboutSection';
-import CitiesSection from '../components/sections/CitiesSection';
+import CitiesSection from '../components/sections/CitySection/CitiesSection';
 import TestimonialsSection from '../components/sections/TestimonialsSection';
 import AchievementsSection from '../components/sections/AchievementSection/AchievementsSection';
-import PartnersSection from '../components/sections/PartnersSection';
+import PartnersSection from '../components/sections/PartnerSection/PartnersSection';
 
 function HomePage() {
   return (
