+++ src/style.css
@@ -88,7 +88,6 @@ main {
     text-align: center;
     padding: 50px 20px;
     background-color: #ffffff;
-    border-bottom: 5px dashed #a0d4b1;
 }
 
 .about-section h2 {
@@ -117,7 +116,6 @@ main {
     text-align: center;
     padding: 50px 20px;
     background-color: #b7e4f7;
-    border-bottom: 5px dashed #a0d4b1;
 }
 
 .cities-section h2 {
@@ -195,8 +193,7 @@ main {
 .testimonials-section {
     text-align: center;
     padding: 50px 20px;
-    background-color: #b7e4f7;
-    border-bottom: 5px dashed #a0d4b1;
+    background-color: white;
 }
 
 .testimonials-section h2 {
@@ -218,7 +215,6 @@ main {
     text-align: center;
     padding: 50px 20px;
     background-color: #f7fffb;
-    border-bottom: 5px dashed #a0d4b1;
 }
 
 .achievements-section h2 {
@@ -236,8 +232,6 @@ main {
 
 .achievement-item {
     background-color: #ffecb3;
-    border: 3px solid #000;
-    border-radius: 20px;
     padding: 20px;
     text-align: center;
     box-shadow: 6px 6px 0px #000;
@@ -247,7 +241,7 @@ main {
 .partners-section {
     text-align: center;
     padding: 50px 20px;
-    background-color: #b7e4f7;
+    background-color:red;
 }
 
 .partners-section h2 {
@@ -266,8 +260,6 @@ main {
 .partner-logo {
     width: 150px;
     height: 100px;
-    border: 3px solid #000;
-    border-radius: 10px;
     background-color: #fff;
     display: flex;
     justify-content: center;
